Restored float and int coordinates rendering.

+++ SISE2015/SISE2015/Globals.cpp
@@ -1,7 +1,7 @@
 #include "Globals.h"
 
-int screen_width;
-int screen_height;
+uint16_t screen_width;
+uint16_t screen_height;
 
 SDL_Window* window;
 SDL_Renderer* renderer;
