+++ SISE2015/SISE2015/Globals.h
@@ -1,13 +1,15 @@
 #ifndef _GLOBALS_H_
 #define _GLOBALS_H_
 
+#include <cstdint>
+
 struct SDL_Window;
 struct SDL_Renderer;
 class DanceMat;
 class Audio;
 
-extern int screen_width;
-extern int screen_height;
+extern uint16_t screen_width;
+extern uint16_t screen_height;
 extern SDL_Window* window;
 extern SDL_Renderer* renderer;
 extern DanceMat* danceMat;
